+++ CHANGELOG.md
@@ -34,7 +34,8 @@ The first release of floodr after converting it from drill. The focus for this r
 - Removed OpenSSL system-level dependency
 - Made exec acitons cross platform
     - Actions now run with `bash` by default on \*nix, `zsh` on macOS and `powershell` on windows. Each will use `-c` with the passed in command as a string, and fail when command returns error status
-    - Added new `--exec-terminal` flag to allow running commands with specified terminal (note `-c "<exec command>"` is passed, make sure your terminal supports this)
+    - Added new `--exec-terminal` flag to allow running commands with specified terminal (note `-c "<exec command>"` is passed, make sure your terminal/language supports this)
+- Delay actions now take `milliseconds` by default instead of `seconds`. If you specify `seconds` it's converted to `miliseconds`. If both are specified `milliseconds` takes precendence
 
 ## Docs